gen-ai/mcp-tutorial/module-03/app/streamlit-app/requirements.txt
streamlit>=1.30.0
strands-agents
strands-agents-tools
mcp>=1.12.2
boto3>=1.34.0
nest-asyncio>=1.6.0